Grid-based A* functional
Units that implement AC_GridNavigation can not pathfind between two points.

AC_GridNavigation FindPath function allows for path determination that is, currently, agnostic of height values and hazards or navigability. It will always return a direct path from point A to point B regardless of walls, etc.

This function has an expansion point to allow for an arbitrary number of additional per-tile heuristics. For example, if a tile reported that it was covered in fire it could notify the A* algorithm that it is a less desirable location to move through.

This function currently has a placeholder neighboring tile retrieval algorithm which will be replaced by the tactical cortex.

M_TileableMaster was moved to a more appropriate folder.

Expanded BPFL_Utils to add a 'pop' function for vector arrays. ToDo: Remake this in C++ to allow for it to function with wildcards.

Updated project-level .gitignore to ignore all files in these folders:
_StarterContent
_Local