The feature
Smooth rotation (turning) with left and right arrow keys and r thumbstick x axis (hopefully, I don't have a gamepad to test it atm 🥲, but it's there) and snap rotations (pivoting) with Q and E (no gamepad input). There's also boolean options to flip the direction of the turning and pivoting individually, perhaps to be exposed to the player.