Bugfixes
Construction actor had a disconnected wire.

BPML math's modulo operation did not account for a modulus of 0 resulting in undefined behavior.

Added an integer calculator-correct modulo operator to BPML math.

BP_PlayerCursor was able to move through walls. It now checks for any X or Y normal that is too close to 1 (absolute) to prevent moving through vertical walls.

AC_GridNavigation's Point Snap To Grid function was buggy. This was due to undefined behavior when rounding a value that is exactly 0.5. It should now operate as intended.