init
established structures and data tables for handling generative grammar tasks.
created  to hold the logic for generative grammar
implemented code to retrieve generative grammar rules from a data table established for that purpose
started work on the logic to find and replace segments of an phrase based on the currently established rules.